Improve validation and error handling
