--- src/services/uploader/types/index.ts
@@ -1,14 +1,0 @@
-export type Upload = {
-  authorization: string;
-  Storage: { Path: string } | undefined;
-  Size: number;
-  ID: string;
-  MetaData: {
-    authorization: string;
-    Storage: any;
-    Size: number;
-    ID: string;
-    upload_id: string;
-    video_id: string;
-  };
-};
